feat: reorganize widget imports and add new components for home view
